requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==12.1.1
gunicorn==21.2.0
bcrypt==4.2.0
Flask-Limiter==3.5.0
